backend/requirements.txt
python-dotenv
phidata
groq
crewai>=0.159.0
crewai-tools==0.62.3
chromadb==0.5.23
langchain
langchain-google-genai
yfinance
duckduckgo-search
serpapi
requests
google-api-python-client
Flask
Flask-CORS
google-cloud-translate
pydub
beautifulsoup4
lxml
scikit-learn
matplotlib
numpy
pandas
google-cloud-texttospeech
google-cloud
google-generativeai
gradio
pinecone
rapidfuzz
packaging
fastapi
uvicorn
openai
pytesseract
opencv-python
Pillow
google-cloud-storage
google-auth
google-auth-oauthlib
google-auth-httplib2